requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==1.0.0
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.12
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
fonttools==4.29.1
grpcio==1.43.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.0
isort==5.10.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
platformdirs==2.5.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
thop==0.0.31.post2005241907
toml==0.10.2
#torch==1.10.2+cpu
#torchvision==0.11.3+cpu
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typed-ast==1.5.2
typing-extensions==4.1.0
urllib3==1.26.8
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.13.3
zipp==3.7.0
python-dotenv==0.19.2